Reworked page gathering to be more efficient and easier to work with
Pages will now properly be tracked and displayed on the ui. Reworked page.cs and pagemanger.cs to now use C# events to track when the player gathers a page. Added PageCountDisplay.cs to now separate the ui element from the Hand.cs script. Hand.cs will now call PageCountDisplay.cs to toggle the page count ui. Added a C# event in the PageManager.cs that fires off when the player gathers all the pages. This event is listened for in the GameManger.cs and plays a victory sound to indicate the player has collected all the pages